3.8 added

+++ ch3/3.8/complex64/main_complex64.go
@@ -23,9 +23,9 @@ func main() {
 
 	img := image.NewRGBA(image.Rect(0, 0, width, height))
 	for py := 0; py < height; py++ {
-		y := float32(py)/height*(ymax-ymin) + ymin
+		y := float64(py)/height*(ymax-ymin) + ymin
 		for px := 0; px < width; px++ {
-			x := float32(px)/width*(xmax-xmin) + xmin
+			x := float64(px)/width*(xmax-xmin) + xmin
 			z := complex(x, y)
 			// Image point (px, py) represents complex value z.
 			img.Set(px, py, mandelbrot(z))
@@ -34,13 +34,13 @@ func main() {
 	png.Encode(os.Stdout, img) // NOTE: ignoring errors
 }
 
-func mandelbrot(z complex64) color.Color {
+func mandelbrot(z complex128) color.Color {
 	const iterations = 200
 	const contrast = 15
 
 	var v complex64
 	for n := uint8(0); n < iterations; n++ {
-		v = v*v + z
+		v = v*v + complex64(z)
 		if cmplx.Abs(complex128(v)) > 2 {
 			return color.Gray{255 - contrast*n}
 		}
